OPEN - issue VISCAT-222: New View for use in Desktop -- non-tabbed Map http://jira.jettmarks.com/jira/browse/VISCAT-222 Lots of refactoring as well

+++ showTrains/src/main/java/com/jettmarks/routes/client/activities/MapActivity.java
@@ -86,7 +86,7 @@ public class MapActivity extends DetailActivity {
 		    public void onTap(TapEvent event) {
 			mapView.clearMap();
 			RouteContainerFactory.getRouteContainer()
-				.setSelectedRoute(null);
+				.setSelectedRoute((Integer) null);
 			clientFactory.getPlaceController()
 				.goTo(new HomePlace());
 		    }
@@ -118,7 +118,7 @@ public class MapActivity extends DetailActivity {
 		    public void onTap(TapEvent event) {
 			mapView.clearMap();
 			RouteContainerFactory.getRouteContainer()
-				.setSelectedRoute(null);
+				.setSelectedRoute((Integer) null);
 			clientFactory.getPlaceController().goTo(
 				new EventSelectionPlace());
 		    }
